split up websocket tests
