+++ CHANGELOG.md
@@ -10,6 +10,7 @@ WebSocket:
 * websocket ping tests
 * Fix websocket close_op resume state
 * websocket write tests
+* split up websocket tests
 
 API Changes:
